# BackendMarketApp/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Precios y Promociones API
  description: API para gestionar precios base y promociones de productos por tienda (usando IDs numéricos incrementales donde aplica), y consultar precios finales.
  version: 1.2.0
servers:
  - url: http://localhost:4000/api/v1
    description: Servidor de Desarrollo Local
...
paths:
...
  /precios/createPrice:
    post:
      tags:
        - Precios
      summary: Crea un nuevo registro de precio histórico.
      description: Añade una nueva entrada de precio para un producto en una tienda específica (usando sus IDs numéricos), válido durante un intervalo de tiempo. Verifica la existencia de la tienda y el producto, y no permite solapamiento con precios existentes para la misma combinación. La BD genera un ID único para este registro de precio.
      operationId: createPrice
      requestBody:
        description: Datos del nuevo precio a crear.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrecioInput'
            examples:
              valid:
                summary: Ejemplo de precio válido
                value:
                  tiendaId: 1
                  productoId: 101
                  valor: 0.95
                  tiempoVigenciaInicio: "2025-07-01T00:00:00Z"
                  tiempoVigenciaFin: "2025-12-31T23:59:59Z"
              invalidDateRange:
                summary: Fechas inválidas (fin antes de inicio o formato incorrecto)
                value:
                  tiendaId: 1
                  productoId: 101
                  valor: 0.95
                  tiempoVigenciaInicio: "2025-08-01T00:00:00Z"
                  tiempoVigenciaFin: "2025-07-31T23:59:59Z"
              invalidValue:
                summary: Valor de precio negativo o no numérico
                value:
                  tiendaId: 1
                  productoId: 101
                  valor: -1.50
                  tiempoVigenciaInicio: "2025-07-01T00:00:00Z"
                  tiempoVigenciaFin: "2025-12-31T23:59:59Z"
              missingField:
                summary: Falta un campo requerido (valor)
                value:
                  tiendaId: 1
                  productoId: 101
                  # valor: 0.95 # Campo faltante
                  tiempoVigenciaInicio: "2025-07-01T00:00:00Z"
                  tiempoVigenciaFin: "2025-12-31T23:59:59Z"
              nonExistentStore:
                 summary: ID de tienda que no existe
                 value:
                   tiendaId: 999 # ID Numérico inválido
                   productoId: 101
                   valor: 0.95
                   tiempoVigenciaInicio: "2025-07-01T00:00:00Z"
                   tiempoVigenciaFin: "2025-12-31T23:59:59Z"
              nonExistentProduct:
                 summary: ID de producto que no existe
                 value:
                   tiendaId: 1
                   productoId: 888 # ID Numérico inválido
                   valor: 0.95
                   tiempoVigenciaInicio: "2025-07-01T00:00:00Z"
                   tiempoVigenciaFin: "2025-12-31T23:59:59Z"
      responses:
        '201':
          description: Precio creado exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Precio creado exitosamente
                  precioRegistroId:
                    type: integer
                    description: El ID numérico autoincremental del registro de precio creado.
                    example: 106
              example:
                message: Precio creado exitosamente
                precioRegistroId: 106
        '400': # Errores de validación de entrada
           description: Solicitud inválida (campos faltantes, formato incorrecto de IDs/valor/fechas, o rango de fechas ilógico).
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/Error'
               examples:
                 missing:
                   value: { message: "Valores faltantes requeridos: valor, ..." }
                 invalidDate:
                   value: { message: "Rango de fecha inválido (formato ISO 8601 esperado y fin debe ser posterior a inicio)." }
                 invalidValue:
                   value: { message: "Valor de precio inválido (debe ser un número no negativo)." }
                 invalidId:
                   value: { message: "ID de tienda inválido (debe ser un número)." }
        '404': # Error de FK no encontrada
           description: No se encontró la Tienda o el Producto especificado.
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/Error'
               examples:
                 storeNotFound:
                   value: { message: "La tienda con ID '999' no existe." }
                 productNotFound:
                   value: { message: "El producto con ID '888' no existe." }
        '409': # Error de solapamiento
          description: Conflicto - El período de precio se solapa con uno existente para la misma tienda y producto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: El rango de fechas se solapa con un precio existente para el mismo producto y tienda.
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /market/queries/final-price:
    get:
      tags:
        - Consultas
      summary: Consulta el precio final de un producto en una tienda en un momento dado.
      description: Calcula el precio final considerando el precio base activo y aplicando la promoción con el mayor porcentaje de descuento que esté activa en la fecha y hora especificadas (usando IDs numéricos para tienda y producto).
      operationId: getFinalPrice
      parameters:
        - name: tiendaId
          in: query
          required: true
          description: El ID numérico de la tienda.
          schema:
            type: integer # ID numérico
          example: 1
        - name: productoId
          in: query
          required: true
          description: El ID numérico del producto.
          schema:
            type: integer # ID numérico
          example: 101
        - name: datetime
          in: query
          required: true
          description: La fecha y hora (en formato ISO 8601 UTC - YYYY-MM-DDTHH:mm:ssZ) para la cual se desea consultar el precio.
          schema:
            type: string
            format: date-time
          example: "2025-07-01T00:00:00Z"
      responses:
        '200':
          description: Precio final calculado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalPriceResponse' # Schema de respuesta
              examples:
                withPromotion:
                  summary: Caso con promoción aplicada
                  value:
                    finalPrice: 0.38
                    basePrice: 0.50
                    appliedPromotionId: 1 # ID numérico de la promoción
                    discountPercentage: 25.00
                    queryDateTime: "2025-07-01T00:00:00Z"
                noPromotion:
                  summary: Caso sin promoción activa
                  value:
                    finalPrice: 0.92
                    basePrice: 0.92
                    appliedPromotionId: null
                    discountPercentage: null
                    queryDateTime: "2025-05-05T08:00:00Z"
                multiplePromotionsBestSelected:
                   summary: Caso con múltiples promociones, se aplica la mejor
                   value:
                     finalPrice: 0.33
                     basePrice: 0.50
                     appliedPromotionId: 2 # ID numérico de la promoción con 35%
                     discountPercentage: 35.00
                     queryDateTime: "2025-05-04T09:00:00Z"
        '400':
          description: Parámetros de consulta inválidos (faltantes o formato incorrecto).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missingParam:
                  value: { message: "Missing required query parameters: datetime." }
                invalidFormat:
                  value: { message: "Invalid datetime format (should be ISO 8601)." }
                invalidId:
                  value: { message: "Invalid tiendaId (must be a number)." }
        '404':
          description: No se encontró un precio base activo para la tienda, producto y hora especificados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: No active base price found for the given product, store, and time.
        '500':
          $ref: '#/components/responses/InternalServerError'
  /market/allStores:
    get:
      tags:
        - Consultas
      summary: Obtiene una lista paginada de tiendas.
      description: Retorna una lista de tiendas registradas, permitiendo paginación a través de los parámetros 'page' y 'limit'. Por defecto, muestra 5 tiendas por página.
      operationId: getAllStores
      parameters:
        - name: page
          in: query
          required: false
          description: Número de la página a obtener. Comienza en 1.
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: limit
          in: query
          required: false
          description: Número máximo de tiendas a retornar por página.
          schema:
            type: integer
            minimum: 1
            maximum: 100 # Puedes definir un máximo razonable
            default: 5
          example: 5
      responses:
        '200':
          description: Lista de tiendas obtenida exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Tiendas obtenidas exitosamente.
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Store' # Referencia al schema de Tienda completo
                  pagination:
                    $ref: '#/components/schemas/Pagination' # Referencia al schema de Paginación
              examples:
                defaultPage: # Ejemplo para la primera página con 5 items
                  summary: Primera página con 5 tiendas
                  value:
                    message: Tiendas obtenidas exitosamente.
                    data:
                      - TiendaID: 1
                        NombreTienda: "Market A"
                      - TiendaID: 2
                        NombreTienda: "Super B"
                      - TiendaID: 3
                        NombreTienda: "Kwik C"
                      - TiendaID: 4
                        NombreTienda: "Tienda D"
                      - TiendaID: 5
                        NombreTienda: "Market Express"
                    pagination:
                      currentPage: 1
                      itemsPerPage: 5
                      totalItems: 23 # Asumiendo 23 tiendas en total
                      totalPages: 5  # 23 / 5 = 4.6 -> 5 páginas
                      hasNextPage: true
                      hasPreviousPage: false
                specificPage: # Ejemplo para la página 2 con límite de 3
                  summary: Página 2 con límite de 3 tiendas
                  value:
                    message: Tiendas obtenidas exitosamente.
                    data:
                      - TiendaID: 4
                        NombreTienda: "Tienda D"
                      - TiendaID: 5
                        NombreTienda: "Market Express"
                      - TiendaID: 6
                        NombreTienda: "MiniMarket Z"
                    pagination:
                      currentPage: 2
                      itemsPerPage: 3
                      totalItems: 23
                      totalPages: 8 # 23 / 3 = 7.6 -> 8 páginas
                      hasNextPage: true
                      hasPreviousPage: true
                emptyResult: # Ejemplo si no hay tiendas o se pide una página fuera de rango
                  summary: Sin resultados (ej. página más allá del total)
                  value:
                    message: Tiendas obtenidas exitosamente.
                    data: []
                    pagination:
                      currentPage: 10
                      itemsPerPage: 5
                      totalItems: 23
                      totalPages: 5
                      hasNextPage: false
                      hasPreviousPage: true # Si totalPages es 5 y pides la 10, la 9 existió
        '400':
          description: Parámetros de consulta 'page' o 'limit' inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Los parámetros \"page\" y \"limit\" deben ser números positivos."
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    PrecioInput: # Schema para crear un precio
      type: object
      required: [tiendaId, productoId, valor, tiempoVigenciaInicio, tiempoVigenciaFin]
      properties:
        tiendaId:
          type: integer
          description: ID numérico de la tienda.
          example: 1
        productoId:
          type: integer
          description: ID numérico del producto.
          example: 101
        valor:
          type: number
          format: float
          description: El valor monetario del precio base.
          example: 0.95
        tiempoVigenciaInicio:
          type: string
          format: date-time
          description: Fecha y hora de inicio de vigencia (ISO 8601 UTC).
          example: "2025-07-01T00:00:00Z"
        tiempoVigenciaFin:
          type: string
          format: date-time
          description: Fecha y hora de fin de vigencia (ISO 8601 UTC). Debe ser posterior al inicio.
          example: "2025-12-31T23:59:59Z"
...
    Store:
      type: object
      properties:
        TiendaID:
          type: integer
          description: ID numérico único de la tienda.
          example: 1
          readOnly: true
        NombreTienda:
          type: string
          description: Nombre de la tienda.
          example: "Market A"
      required:
        - TiendaID
        - NombreTienda
    Pagination: # Schema para la metadata de paginación
      type: object
      properties:
        currentPage:
          type: integer
          description: La página actual de resultados.
          example: 1
        itemsPerPage:
          type: integer
          description: El número de items solicitados por página.
          example: 5
        totalItems:
          type: integer
          description: El número total de items disponibles.
          example: 23
        totalPages:
          type: integer
          description: El número total de páginas disponibles.
          example: 5
        hasNextPage:
          type: boolean
          description: Indica si hay una página siguiente.
        hasPreviousPage:
          type: boolean
          description: Indica si hay una página anterior.
      required:
        - currentPage
        - itemsPerPage
        - totalItems
        - totalPages
        - hasNextPage
        - hasPreviousPage

    # --- Schemas de Respuesta ---
    FinalPriceResponse: # Schema para la respuesta de la consulta
      type: object
      properties:
        finalPrice:
          type: number
          format: float
          description: El precio final calculado después de aplicar el mejor descuento.
        basePrice:
          type: number
          format: float
          description: El precio base encontrado para la fecha/hora.
        appliedPromotionId:
          type: integer
          nullable: true
          description: El ID numérico de la promoción que se aplicó (o null).
        discountPercentage:
          type: number
          format: float
          nullable: true
          description: El porcentaje de descuento aplicado (o null).
        queryDateTime:
          type: string
          format: date-time
          description: La fecha/hora para la que se realizó la consulta.
    Error: # Schema genérico de error
      type: object
      properties:
        message:
          type: string
          description: Descripción del error ocurrido.
      required:
        - message
  responses: # Definiciones genéricas de respuestas
...
    InternalServerError:
      description: Error interno del servidor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Internal Server Error"